refactor: dummyData변경, 주석삭제, 사용하지 않은 코드 삭제
